fix: move calc_test() to src/calc/test.c
